+++ src/MicroElements.Metadata.Xml/Xml/XmlParserSettings.cs
@@ -27,6 +27,9 @@ namespace MicroElements.Metadata.Xml
         public IPropertyValueFactory PropertyValueFactory { get; }
 
         /// <inheritdoc />
+        public IStringProvider StringProvider { get; }
+
+        /// <inheritdoc />
         public IValidationProvider ValidationProvider { get; }
 
         /// <inheritdoc />
@@ -37,6 +40,7 @@ namespace MicroElements.Metadata.Xml
         /// </summary>
         /// <param name="getElementName">Function that evaluates property name for xml element.</param>
         /// <param name="parserRules">Parsers and rules for parsers.</param>
+        /// <param name="stringProvider">Optional string provider.</param>
         /// <param name="propertyComparer">Property comparer for property related search. Default value: <see cref="Metadata.PropertyComparer.ByReferenceComparer"/>.</param>
         /// <param name="propertyValueFactory"><see cref="IPropertyValue"/> factory. Default: <see cref="CachedPropertyValueFactory"/>.</param>
         /// <param name="validationFactory">Validation factory to create validation rules.</param>
@@ -44,6 +48,7 @@ namespace MicroElements.Metadata.Xml
         public XmlParserSettings(
             Func<XElement, string>? getElementName = null,
             IReadOnlyCollection<IParserRule>? parserRules = null,
+            IStringProvider? stringProvider = null,
             IEqualityComparer<IProperty>? propertyComparer = null,
             IPropertyValueFactory? propertyValueFactory = null,
             IValidationProvider? validationFactory = null,
@@ -51,6 +56,7 @@ namespace MicroElements.Metadata.Xml
         {
             GetElementName = getElementName ?? XmlParser.GetElementNameDefault;
             ParserRules = parserRules ?? XmlParser.CreateDefaultXmlParsersRules().ToArray();
+            StringProvider = stringProvider ?? new DefaultStringProvider();
             PropertyComparer = propertyComparer ?? Metadata.PropertyComparer.ByReferenceComparer;
             PropertyValueFactory = propertyValueFactory ?? new CachedPropertyValueFactory(new PropertyValueFactory(), PropertyComparer);