+++ src/MicroElements.Metadata/Metadata/ParserRuleExtensions.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using MicroElements.Metadata.Parsers;
+using MicroElements.Metadata.Parsing;
 
 namespace MicroElements.Metadata
 {