+++ src/MicroElements.Metadata/Metadata/Parsing/Parser.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Globalization;
-using MicroElements.Metadata.Parsers;
 
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member